Fix: hoc.nextedComment -> Root.comment로 변경
