+++ webapp/src/components/ComentContainer/CommentList/Root.CommentList.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { useCommentsAction, useCommentsState } from 'contexts/Comment/Comment.Provider';
 import HocSecretComment from '../Hoc/HocSecretComment';
-import HocNestedComment from '../Hoc/Hoc.NestedComment';
+import RootCommentElement from '../CommentElement/Root.CommentElement';
 import * as S from '../style';
 
 RootCommentList.propTypes = {
@@ -21,7 +21,7 @@ export default function RootCommentList({ postWriter, comments }) {
       {comments.map(({ id, teamId, userId, replies, ...commentInfo }) => {
         const { secret, writer: commenWriter } = commentInfo;
         const isSecret = isShowSecretComment(secret, postWriter, commenWriter, loggedInUserName);
-        const WithSecretComment = HocSecretComment({ isSecret, Component: HocNestedComment });
+        const WithSecretComment = HocSecretComment({ isSecret, Component: RootCommentElement });
         return (
           <WithSecretComment
             key={id}
